get rid of gtest and gmock in ITF
