+++ test/integration/pipeline/tx_pipeline_integration_test_fixture.hpp
@@ -24,72 +24,15 @@
 #include "cryptography/ed25519_sha3_impl/internal/sha3_hash.hpp"
 #include "datetime/time.hpp"
 #include "framework/test_subscriber.hpp"
-#include "main/application.hpp"
 #include "main/raw_block_insertion.hpp"
 #include "model/generators/block_generator.hpp"
 #include "module/irohad/ametsuchi/ametsuchi_fixture.hpp"
+#include "test_irohad.hpp"
 
 using namespace framework::test_subscriber;
 using namespace std::chrono_literals;
 using namespace iroha::model::generators;
 using iroha::model::Transaction;
-
-class TestIrohad : public Irohad {
- public:
-  TestIrohad(const std::string &block_store_dir,
-             const std::string &redis_host,
-             size_t redis_port,
-             const std::string &pg_conn,
-             size_t torii_port,
-             size_t internal_port,
-             size_t max_proposal_size,
-             std::chrono::milliseconds proposal_delay,
-             std::chrono::milliseconds vote_delay,
-             std::chrono::milliseconds load_delay,
-             const iroha::keypair_t &keypair)
-      : Irohad(block_store_dir,
-               redis_host,
-               redis_port,
-               pg_conn,
-               torii_port,
-               internal_port,
-               max_proposal_size,
-               proposal_delay,
-               vote_delay,
-               load_delay,
-               keypair) {}
-
-  auto &getCommandService() {
-    return command_service;
-  }
-
-  auto &getQueryService() {
-    return query_service;
-  }
-
-  auto &getPeerCommunicationService() {
-    return pcs;
-  }
-
-  auto &getCryptoProvider() {
-    return crypto_verifier;
-  }
-
-  void run() override {
-    grpc::ServerBuilder builder;
-    int port = 0;
-    builder.AddListeningPort("0.0.0.0:" + std::to_string(internal_port_),
-                             grpc::InsecureServerCredentials(),
-                             &port);
-    builder.RegisterService(ordering_init.ordering_gate_transport.get());
-    builder.RegisterService(ordering_init.ordering_service_transport.get());
-    builder.RegisterService(yac_init.consensus_network.get());
-    builder.RegisterService(loader_init.service.get());
-    internal_server = builder.BuildAndStart();
-    internal_thread = std::thread([this] { internal_server->Wait(); });
-    log_->info("===> iroha initialized");
-  }
-};
 
 class TxPipelineIntegrationTestFixture
     : public iroha::ametsuchi::AmetsuchiTest {
